requirements.txt
defusedxml==0.5.0
Django==1.10.1
djangorestframework==3.6.2
easy-thumbnails==2.4.1
mysqlclient==1.3.7
oauthlib==2.0.0
olefile==0.44
Pillow==4.1.0
Pygments==2.2.0
PyJWT==1.4.2
python-social-auth==0.2.21
python3-openid==3.0.10
PyYAML==3.12
requests==2.11.1
requests-oauthlib==0.6.2
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
sorl-thumbnail==12.3
virtualenv==15.1.0
